update serializers with error_details
