+++ app/controllers/api/v1/credit_notes_controller.rb
@@ -31,7 +31,7 @@ module Api
           json: ::V1::CreditNoteSerializer.new(
             credit_note,
             root_name: 'credit_note',
-            includes: %i[items applied_taxes]
+            includes: %i[items applied_taxes error_details]
           )
         )
       end
@@ -109,7 +109,7 @@ module Api
             ::V1::CreditNoteSerializer,
             collection_name: 'credit_notes',
             meta: pagination_metadata(credit_notes),
-            includes: %i[items applied_taxes]
+            includes: %i[items applied_taxes error_details]
           )
         )
       end
