+++ app/graphql/types/credit_notes/object.rb
@@ -47,6 +47,7 @@ module Types
 
       field :external_integration_id, String, null: true
       field :integration_syncable, GraphQL::Types::Boolean, null: false
+      field :error_details, [Types::ErrorDetails::Object], null: true
 
       def applied_taxes
         object.applied_taxes.order(tax_rate: :desc)
