+++ app/serializers/v1/error_detail_serializer.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module V1
+  class ErrorDetailSerializer < ModelSerializer
+    def serialize
+      {
+        lago_id: model.id,
+        error_code: model.error_code,
+        details: model.details
+      }
+    end
+  end
+end
