--- app/serializers/v1/invoices/error_detail_serializer.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module V1
-  module Invoices
-    class ErrorDetailSerializer < ModelSerializer
-      def serialize
-        {
-          lago_id: model.id,
-          error_code: model.error_code,
-          details: model.details
-        }
-      end
-    end
-  end
-end
